# Avoid auto-diff for linear IPOPT constraints