* Addresses the following IPOPT issue mentioned in [#18292](https://github.com/RobotLocomotion/drake/issues/18292#issuecomment-1314121771) by avoiding auto-diff for linear constraints:
> We treat the linear constraints as if they are nonlinear. When we compute the gradient of the linear constraints lb <= A * x <= ub, instead of fetching the A matrix directly, we form an AutoDiffVecXd, evaluate this constraint with AutoDiffVecXd, and then take the gradient. This is very inefficient. We should directly fetch A matrix as the gradient without evaluating AutoDiff. This is a problem specific in IpoptSolver, but not in SnoptSolver.

* Adds a benchmark to evaluate the IPOPT performance for a nonlinear program in the following form:
`min x^T*x s.t. x^T*x <= 1, Identity*x  = 0, 0 <= A*x <= 1e20, -1 <= x <= 1`
where A is a random matrix.

The benchmark results show that the solver speed is improved significantly, especially for large-enough problems, by the proposed change.

| Version | Size | Time [ns] | CPU [ns] | Iters |
| ----------- | ----------- | ----------- | ----------- | ----------- |
| Baseline    | 10 |  1226912 | 1226656   | 555 |
| Proposed  | 10 |  1200639 |  1200614  | 581 |
| Baseline    | 100 |  15413920 |  15413016  | 45 |
| Proposed  | 100 |  13354838 |  13353761  | 52 |
| Baseline    | 1000 | 9818495035  |  9817708665  | 1 |
| Proposed  | 1000 | 7225218296  |  7224712225  | 1 |

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RobotLocomotion/drake/18603)
<!-- Reviewable:end -->